Refactor project structure and import paths
- Reorganize project files into more standard directory structure
- Update import paths to use lowercase 'hooks' and 'pages' directories
- Remove unused useRegister.js hook
- Fix minor typos and syntax errors in components
- Standardize import statements across project files
