+++ src/hooks/useContext.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useContext } from "react";
-import { ChatContent } from "../Component/ChatProvider";
+import { ChatContent } from "../pages/ChatProvider";
 // import { ChatContent } from "./ChatProvider";
 
 export const useContextChat = () => {
